[Feature] Add iOS renderer host support and NativePropBundle integration
- What changed
  - Added renderer host registration APIs (LYNX_REGISTER_RENDERER_HOST, LYNX_LAZY_REGISTER_RENDERER_HOST) to LynxComponentRegistry, enabling custom renderer hosts for PlatformRendererDarwin.
  - Implemented PlatformRendererDarwin initialization with custom renderer hosts via LynxComponentRegistry.
  - Integrated NativePropBundle into PropBundleDarwin and updated PropBundleCreatorDarwin to support native value creation.
  - Refactored LynxRendererHost protocol to use properties (renderer, rendererContext, view) and added initWithRendererContext:.
  - Renamed LynxRenderer accessor methods (getSign -> sign, getRenderer -> renderer, getView -> view) and added updateAttributes: / updatePlatformExtraBundle: hooks.
  - Implemented GetTagInfo and UpdatePlatformExtraBundle in native painting context for iOS.
  - Updated layout context to handle NativePropBundle when creating/updating layout nodes.
  - Fixed a typo in LynxComponentScopeRegistry (makeIntoGloabl -> makeIntoGlobal).
  - Fixed implicit cast warnings in NativePropBundle::SetPropsByID.
  - Updated lynx_ios.api to reflect all API changes.
  - Added necessary CoreGraphics import to LynxConverter.h.

- Why it was needed
  - To support custom renderer hosts on iOS, allowing extended renderers to be backed by user-provided host views.
  - To unify prop bundle handling across platforms by leveraging NativePropBundle on iOS.

- How it was verified
  - Existing LynxDisplayListApplierUnitTest was updated for the new LynxRendererHost protocol.
  - API file lynx_ios.api was synchronized with the new public interfaces.

SkipChecks: macro
AutoSubmit: true

+++ core/renderer/ui_wrapper/common/native_prop_bundle.cc
@@ -101,7 +101,7 @@ void NativePropBundle::SetPropsByID(CSSPropertyID id, const uint8_t* data,
                                     size_t size) {
   auto array = lepus::Value(lepus::CArray::Create());
   for (size_t i = 0; i < size; ++i) {
-    array.SetProperty(i, lepus::Value(data[i]));
+    array.SetProperty(static_cast<uint32_t>(i), lepus::Value(data[i]));
   }
   auto key = CSSProperty::GetPropertyNameCStr(id);
   props_[key] = lepus::Value(std::move(array));
@@ -111,7 +111,7 @@ void NativePropBundle::SetPropsByID(CSSPropertyID id, const uint32_t* data,
                                     size_t size) {
   auto array = lepus::Value(lepus::CArray::Create());
   for (size_t i = 0; i < size; ++i) {
-    array.SetProperty(i, lepus::Value(data[i]));
+    array.SetProperty(static_cast<uint32_t>(i), lepus::Value(data[i]));
   }
   auto key = CSSProperty::GetPropertyNameCStr(id);
   props_[key] = lepus::Value(std::move(array));
